Create your own test

This page illustrates how you can create univariate and multiple comparisons permutation tests besides those already supported by PermutationsTests.jl.

Good knowledge and understanding of the whole documentation is required for continuing reading this page, including the Extras pages.

There are two ways you can extend the capability of PermutationTests.jl.

The first, very simple, but limited, is by using existing tests with particular data input.

The second, which needs some development, but is much more general, allows you to create new tests defining your own test statistics.

Using existing tests

The approach is illustrated by creating tests for the autocorrelation, which are not in the API of the package.

Example of a univariate permutation test

As an example, let us detect the presence of autocorrelation at lag 1, similar to what the Durbin-Watson test does.

PermutationTests.jl implements the Pearson product moment correlation statistic and a test for it. We can then test the null hypothesis that the autocorrelation at lag 1 is equal to zero by performing a correlation test on a data vector x and a lagged version of itself y.

This is simply achieved by calling the _permTest!, which performs all univariate tests implemented in the package:

using PermutationTests
# First, let us create some data autorrelated at lag 1 as example
N=300
data=randn(N)
for i=2:length(data)
  data[i-1]=data[i-1]+data[i]
end

# Data and lagged version of the data
x=data[1:end-1]
y=data[2:end]

# Our Univariate test on autocorrelation
t=_permTest!(copy(x), y, N-1, PearsonR(), PearsonR())
# NB: we pass a copy of x as this vector will be permuted.

⚅ Univariate Permutation test
p-value = <0.001
⚁      ⚁      ⚂      ⚃      ⚁      ⚃      ⚄      ⚁      ⚄      ⚀      ⚀      
.p (p-value) .stat (test statistic).obsstat   (observed statistic)
.minp  (minimum attainable p-value).nperm (number of permutations)
.testtype  (exact or approximated).direction (Both, Left or Right)
.design  (Balanced or Unbalanced)

The testing procedure can be further improved and refined, see for example correlationTest, but this should suffice to give the idea.

Example of a multiple comparison permutation test

Similarly, we can easily test simultaneously the autocorrelation at several lags with a multiple comparison version of the above test.

This is achieved by calling the _permMcTest!, which performs all multiple comparisons tests implemented in the package. Now we craete a vector x and a vector Y holding as many lagged versions of x as we wish to test:

using PermutationTests
# Let us create some data autorrelated at lag 1 and 2 as example
N=300
data=randn(N)
for j=1:2, i=2:length(data)
  data[i-1]=data[i-1]+data[i]
end

# We will test lags 1 to 30
lags=1:30
x=data[1:end-length(lags)]
Y=[data[l+1:end-length(lags)+l] for l in lags]

# Our muliple comparisons test on autocorrelations
tMc=_permMcTest!(copy(x), Y, N-length(lags), PearsonR(), PearsonR())

⚀ ⚀ ⚂ ... Multiple Comparison Permutation test
Rejected 2 out of 30 hypotheses (6.67 %) in 1 step with FWE=0.05
⚅      ⚄      ⚂      ⚂      ⚂      ⚄      ⚄      ⚂      ⚄      ⚁      ⚀      
.p (p-values).stat (test statistic).obsstat  (observed statistics)
.minp  (minimum attainable p-value).nperm (number of permutations)
.testtype  (exact or approximated).direction (Both, Left or Right)
.design  (Balanced or Unbalanced)
.nulldistr   (null distribution).rejections (for each step down)
.stepdown    (true if stepdown).fwe (family-wise err. stepdown)

The p-values are retrived by

tMc.p # p-values for lags 1...30 

Notice that the test above is different from the Liung-Box test, which is a multivariate test for a group of lags: with a multiple comparison tests we have tested each lag in the group simultaneously, controlling the family-wise error (FWE) rate at the nominal level (0.05 by default). Thus on the base of the p-values we can take a decision on the null hypothesis at each lag.

You can check yourself that the test controls the FWE rate at the nominal level by running the test above with random Gaussian data many times and controlling that the proportion of rejections, which are false by definition since random Gaussian data is not autocorrelated, does not exceed the nominal level.

Defining new test statistics

You can create your own permutation tests based on a custom test statistic as long as the permutation scheme for your test is supported by PermutationTests.jl. Four permutation schemes are implemented in the package, corresponding to four implemented groups of test-statistics. Before continuing, read the documentation of genPerms and the documentation linked therein.

In order to create a test you will call function _permTest!(univariate) or _permMcTest!(multiple comaprisons). You will pass, among others, argument asStat, which determines the permutation scheme of the test.

keep in mind

The asStat argument can be one of the following built-in test statistics or any test statistic belonging to the same group:

PearsonR()

AnovaF_IS()

AnovaF_RM()

StudentT_1S()

The procedure to follow is slightly different depending on whether you want to create your own univariate or multiple comparisons test.

UNIVARIATE TESTS

Creating your own univariate test involves two preliminary steps:

define a new Statistic type to name the test statistic you want to use

write a method for the statistic function to compute the test statistic for both the observed and permuted data

You then obtain the test just calling the _permTest! functions, as we will show.

keep in mind

Argument fstat of _permTest! by default is the abs julia function, which yields a bi-directional test. If appropriate for the test statistic at hand, use fstat=identity for a right-directional test and fstat=flip for a left-directional test.

Example of a univariate permutation test

We create a new test for the difference of the mean between two independent samples. Such a test is equivalent to the Student's t-test for independent samples, since the mean difference is the nominator of the t statistic and the denominator is invariant with respect to data permutation.

The permutations scheme is the same as the permutation scheme of the AnovaF_IS() (or StudentT_IS) test statistic, hence it is supported by PermutationTests.jl and we can reuse it.

# First, let us import `statistic` from PermutationTests.jl
using PermutationTests
import PermutationTests: statistic

# 1) define a new `Statistic` type to name the test statistic you want to use
struct MeanDiff <: Statistic end

#=
2) write a method for the `statistic` function to compute the mean difference,
it does not matter if the data is observed or permuted.
We must input the data exactly as for the `AnovaF_IS()` statistic.

Our observations are divided in two groups. In this case,
ror this and equivalent statistics, data input is given as a single vector
y of length N1+N2, with the N1 observations followed by the N2 observations.

The `statistic` method also takes as input a 'membership' vector x,
holding 1 for subjects of group 1 and 2 for subjects of group 2,
that is, x=[repeat ([1], N1); repeat([2], N2)].
When given as input, x will reflect the positioning of the observed data
and for permuted data it will be a permutation thereof.

A simple implementation:
=#
function statistic(x, y, stat::MeanDiff; kwargs...)
    # get the number of subjects in each gruoup
    ns=[count(j->j==i, x) for i=1:length(unique(x))]

    s=[0., 0.]
    @simd for i ∈ eachindex(x, y)
        @inbounds (s[x[i]] += y[i])
    end
    return (s[1]/ns[1]) - (s[2]/ns[2]) # mean difference
end

#=
The following code block is not necessary.
We report it to illustrate a nice opportunity you have to write
efficient code to compute test statistics;
the above implementation computes the group numerosity (ns=[N1, N2])
at each permutation, however ns is invariant to permutations.

Let us rewrite the function then using the `cpcd` argument,
which will be passed to the `_permTest!` functions in order to
(optionally) pre-compute `ns`.
=#
function statistic(x, y, stat::MeanDiff; cpcd=nothing, kwargs...)
    if cpcd===nothing
        ns=[count(j->j==i, x) for i=1:length(unique(x))]
    else
        ns=cpcd
    end

    s=[0., 0.]
    @simd for i ∈ eachindex(x, y)
        @inbounds (s[x[i]] += y[i])
    end
    return (s[1]/ns[1]) - (s[2]/ns[2]) # mean difference
end

# That's it. Let's test our own test:

# create some data
y=[4., 5., 6., 7., 1., 2., 3]
ns=[4, 3] # N1=4, N2=3

# membership vector (the dedicated function `membership` is available)
x=membership(StudentT_IS(), ns) # = [repeat ([1], N1); repeat([2], N2)]

println("t-test independent samples (bi-directional) using PermutationsTest.jl: ");
# two-directional exact test. Equivalent to API function studentTestIS(y, ns)
t1 = _permTest!(copy(x), y, ns, StudentT_IS(), StudentT_IS())
# Note: we pass a copy of x as this vector will be permuted.

# This runs the test we have created. Note that argument `stat` of `_permTest!`
# is `MeanDiff()`, while argument `asStat` is `AnovaF_IS()`
println("t-test independent samples (bi-directional) using our own statistic: ");
t2 = _permTest!(copy(x), y, ns, MeanDiff(), AnovaF_IS())

⚃ Univariate Permutation test
p-value = 0.057
⚂      ⚂      ⚁      ⚀      ⚂      ⚁      ⚁      ⚅      ⚀      ⚀      ⚃      
.p (p-value) .stat (test statistic).obsstat   (observed statistic)
.minp  (minimum attainable p-value).nperm (number of permutations)
.testtype  (exact or approximated).direction (Both, Left or Right)
.design  (Balanced or Unbalanced)

As before, but using the cpcd argument to avoid redundant computations:

t2_ = _permTest!(copy(x), y, ns, MeanDiff(), StudentT_IS(); cpcd=ns) 

You can check that the p-values obtained by the two tests is identical:

t1.p≈t2.p ? (@info "my test works!") : 
            (@error "my test does not work!" t1.p t2.p)

To run a right-directional test:

t3 = _permTest!(copy(x), y, ns, MeanDiff(), StudentT_IS(); fstat = identity)

To run a left-directional approximate test (see _permTest!):

t4 = _permTest!(copy(x), y, ns, MeanDiff(), StudentT_IS(); fstat = flip, switch2rand=1)

etc.

MULTIPLE COMPARISON TESTS

Creating your own multiple comparison test involves three preliminary steps:

define a new Statistic type to name the test statistic you want to use

write a method of the _observedStats function to compute function fstat of all observed test-statistics at once as a vector

write a method of the _permutedStat function to compute function fstat of each permutation statistics separately

You then obtain the test just calling the _permMcTest! functions.

keep in mind

Argument fstat of _permMcTest! by default is the abs julia function, which yields a bi-directional test. If appropriate for the test statistic at hand, use fstat=identity for a right-directional test and fstat=flip for a left-directional test.

Example of a multiple comparison permutation test

We create another version of the Pearson product-moment correlation test between a fixed variable given as a vector x and $M$ variables given as a vector of $M$ vectors Y. We want to test simultaneously all correlations between x and Y[m], for $m=1...M$.

The permutations scheme is the one of the PearsonR() statistic, hence it is supported by PermutationTests.jl and we can reuse it (as asStat argument of _permMcTest!).

# First, let us import some names from PermutationTests.jl
using PermutationTests
import PermutationTests: statistic, _observedStats, _permutedStat

using LinearAlgebra: dot, ⋅ # we are going to use the dot product

# 1) Define a new `Statistic` type to name the test statistic you want to use
struct MyPearsonR <: Statistic end

#=
2) Write a function to compute function `fstat` of all observed
test-statistics at once as a vector.

In order to obtain a faster test we will work with standardized
variables, thus the Pearson correlation is given simply by the
cross-product divided by N.

Note that we will subtract a small constant to all
obseved statistics (sqrt(eps())) to avoid numerical errors when
comparing the observed statistics to the permuted statistics.
=#
_observedStats(x, Y, stat::MyPearsonR, fstat::Function=abs; kwargs...) =
    [fstat((x ⋅ y)/length(x)) - sqrt(eps()) for y ∈ Y]

#=
3) Write a function to compute function `fstat` of the mth
permuted test-statistic.

Again, we will work with standardized variables, thus the Pearson
correlation is given simply by the cross-product divided by N.
=#
_permutedStat(x, Y, m::Int, stat::MyPearsonR, fstat::Function=abs; kwargs...) =
    fstat((x ⋅ Y[m])/length(x))

# That's it. Let's test our own test:

# Let's create some data as example
N, M = 7, 100 # N=7 observations, M=100 hypotheses
x=randn(N);
Y=[randn(N) for m=1:M];

println("correlation test using PermutationsTest.jl: ");
# Exact test. Equivalent to API function rMcTest!(x, Y)
# NB: the standardization function μ0σ1() is exported by PermutationTests.jl
t1 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y ∈ Y], length(x), PearsonR(), PearsonR())

println("correlation test using our own test: ");
t2 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y ∈ Y], length(x), MyPearsonR(), PearsonR())

⚁ ⚃ ⚂ ... Multiple Comparison Permutation test
Rejected 0 out of 100 hypotheses (0.0 %) in 1 step with FWE=0.05
⚄      ⚄      ⚀      ⚀      ⚃      ⚅      ⚂      ⚅      ⚁      ⚃      ⚃      
.p (p-values).stat (test statistic).obsstat  (observed statistics)
.minp  (minimum attainable p-value).nperm (number of permutations)
.testtype  (exact or approximated).direction (Both, Left or Right)
.design  (Balanced or Unbalanced)
.nulldistr   (null distribution).rejections (for each step down)
.stepdown    (true if stepdown).fwe (family-wise err. stepdown)

Check that the p-values obtained by the two tests is identical:

abs(sum(t1.p-t2.p))<1e-8 ?  (@info "my test works!") : 
                            (@error "my test does not work!")

Check also the observed statistics:

abs(sum(t1.obsstat-t2.obsstat))<1e-8 ?  (@info "my test works!") : 
                                        (@error "my test does not work!")

To run a right-directional test:

t3 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y ∈ Y], length(x), MyPearsonR(), PearsonR(); 
        fstat = identity) 

To run a left-directional approximate test:

t4 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y ∈ Y], length(x), MyPearsonR(), PearsonR(); 
        fstat = identity, switch2rand=1)

We can also pre-compute data that is invariant to permutations in order to obtain a faster procedure.

For example, suppose we didn't know that the cross-product divided by N is equal to the Pearson statistic when the data is standardized. Instead of computing the Pearson statistics each time from scratch at each permutation, we could still pre-compute the means and standard deviations, which are invariant to pemutations, using the cpcd (custom pre-computed data) keyword argument, such as:

# Pre-compute means and standard deviations. Here you can compute anything you need
# NB: the mean μ() and st. deviation function σ(y) are exported by PermutationTests.jl 
cpcd=[μ(x), [μ(y) for y in Y], σ(x), [σ(y) for y in Y]]

Declare functions for observed and permuted data using argument cpcd. Note that

_observedStats computes all $M$ statistics, while _permutedStat computes the $m^{th}$ statistic.

not the statistics, but function fstat of the statistics are returned

_observedStats(x, Y, stat::MyPearsonR, fstat::Function=abs; cpcd=cpcd, kwargs...) = 
    [fstat( (((x.-cpcd[1])./cpcd[3]) ⋅ ((y.-cpcd[2][m])./cpcd[4][m])) / length(x) ) - 
        sqrt(eps()) for (m, y) ∈ enumerate(Y)]

_permutedStat(x, Y, m::Int, stat::MyPearsonR, fstat::Function=abs; cpcd=cpcd, kwargs...) = 
    fstat((((x.-cpcd[1])./cpcd[3]) ⋅ ((Y[m].-cpcd[2][m])./cpcd[4][m])) / length(x))

println("correlation test using my own test and pre-computed data: ");
t2_ = _permMcTest!(copy(x), Y, length(x), MyPearsonR(), PearsonR(); cpcd=cpcd) 

Note that we copy x because otherwise, using this permutation scheme, it is overwritten. If you don't care, you can simply pass x.

nota bene

If you reuse the permutations scheme given with asStat=StudentT_1S(), x or Y are overwritten, depending on whether the test is exact or approximated.

Note also that now running the _permMcTest! function without the cpcd kwarg gives an error, since we did not implement versions of _observedStats and of _permutedStat that take it as an optional argument (although we could, as we have done for the univariate test).

PermutationTests._permTest! — Function

function _permTest!(x::UniData, y::UniData, ns::nsType, stat::Stat, asStat::AsStat;
                    standardized::Bool=false, centered::Bool=false, 
                    nperm::Int = 20000, 
                    fstat::Function = abs,
                    compfunc::Function = >=,
                    switch2rand::Int = Int(1e8),
                    seed::Int = 1234,
                    verbose::Bool = true,
                    cpcd = nothing) where {Stat<:Statistic, AsStat<:Statistic}

This function ultimately performs all univariate permutation tests implemented in PermutationsTests.jl, both exact and approximate (Monte Carlo).

For running tests use the univariate test functions. You need this function only for creating your own tests.

Rewrite x and/or y, depending on the test performed.

For the ns argument see ns.

stat can be a singleton of the Statistic type or a user-defined singleton of this type to be used as argument of a function implemented by the user to compute both the observed and permuted test statistics, see create your own test.

asStat is a singleton of the Statistic type. It is used to determine the permutation scheme and for this purpose it will be passed to genPerms and nrPerms.

asStat determines also the input data format if you declare your own stat type. In this case the function you write for computing the observed and permuted test statistic will take the x and y arguments as it follows:

For Stat belonging to group

BivStatistic : x, y are the two vectors of $N$ elements each for which the bivariate statistic is to be tested. ns is ignored.

IndSampStatistic : we have $K$ groups and $N=N_1+...+N_K$ total observations; y holds all observations in a single vector such as [y1;...;yK] and x is the membership(::IndSampStatistic) vector. For example, for $K=2$, $N_1=2$ and $N_2=3$, x=[1, 1, 2, 2, 2].

RepMeasStatistic : we have $K$ measures (e.g., treatements) and $N$ subjects; y holds the $K*N$ observations in a single vector such as [y1;...;yN], where each vector $y_i$, for $i=1...N$, holds the observation at the $K$ treatments and x=collect(1:K*N) (see membership(::RepMeasStatistic)).

OneSampStatistic : We have $N$ observations (e.g., subjects); y holds the $N$ observations and x=ones(Int, N) (see membership(::OneSampStatistic)).

For length(x)>30 the approximate test is performed in all cases,

otherwise,

if the number of systematic permutations exceeds switch2rand the approximate test is performed using nperm random permutations (default 20000),

otherwise,

the exact test is performed.

switch2rand defaults to 1e8. To perform approximate tests in all cases, set switch2rand, for example, to 1.

If stat is a BivStatistic, it optionally uses kwargs standardized or centered to compute them faster, see for example correlationTest.

seed is the initial seed for generating random permutations (not used for exact tests). To use a random seed, pass seed=0. For seed any natural number, the test will be reproducible.

fstat is a function applied to the test statistic. By default this is the julia abs function, which takes the absolute value, hence yieds a bi-directional test. For a right-directional test pass here identity. For a left-directional test pass here flip.

compfunc is the function to compare the observed statistics to the permuted statistics. The default function is >=. Don't change it.

If verbose is true, print some information in the REPL while running the test. Set to false if you run benchmarks. The default is true.

For the cpcd argument, see create you own test.

Return a UniTest structure.

Examples

using PermutationTests
x=randn(6)
y=randn(6) 
# bi-directional exact test of the correlation between x and y
t8 = _permTest!(μ0(x), μ0(y), length(x), Covariance(), Covariance(); centered=true) 
t8.p
t8.stat
#...

# make a left-directional test and standardize the data
t8_2 = _permTest!(μ0σ1(x), μ0σ1(y), length(x), Covariance(), Covariance(); 
        standardized=true, fstat=flip) 

# the same but force an approximate test
t8_3 = _permTest!(μ0σ1(x), μ0σ1(y), length(x), Covariance(), Covariance(); 
        standardized=true, fstat=flip, switch2rand=1) 

# the same using 5000 random permutations
t8_4 = _permTest!(μ0σ1(x), μ0σ1(y), length(x), Covariance(), Covariance(); 
        standardized=true, fstat=flip, switch2rand=1, nperm=5000) 

To check more examples, see the uniTests_API.jl unit located in the src github folder and function test_unitests() in the runtests.jl unit located in the test github folder.

PermutationTests._permMcTest! — Function

function _permMcTest!(x::UniData, Y::UniDataVec, ns::nsType, stat::Stat, asStat::AsStat;
            standardized::Bool=false, centered::Bool=false, 
            stepdown::Bool = true,
            fwe::Float64 = 0.05,
            nperm::Int = 20000, 
            fstat::Function = abs,
            compfunc::Function = >=,
            switch2rand::Int = Int(1e8),
            seed::Int = 1234,
            threaded::Bool = Threads.nthreads()>=4,
            verbose::Bool = true,
            cpcd = nothing) where {Stat<:Statistic, AsStat<:Statistic}

This function ultimately performs all multiple comparisons permutation tests implemented in PermutationsTests.jl, both exact and approximate (Monte Carlo).

For running tests use the multiple comparisons test functions. You need this function only for creating your own tests.

The step-down version of the test is performed if stepdown is true (default). In this case the fwe (family-wise error) rate is used for rejection at each step (default=0.05).

Rewrite x and/or Y, depending on the test performed.

For the ns argument see ns.

stat can be a singleton of the Statistic type or a user-defined singleton of this type to be used as argument of two functions implemented by the user to compute the observed and permuted test statistics, see create your own test.

Warning

In contrast to univariate tests, equivalent statistics are not possible for multiple comparison tests, with the exception of CrossProd() if the data has been standardized and Covariance() if the data has been centered and those only for correlation-like tests.

The test statistics that must be used as Stat for the other kinds of test if a singleton of the Statistic type is used are AnovaF_IS(), StudentT_IS(), AnovaF_RM(), and StudentT_1S().

asStat is a singleton of the Statistic type. It is used to determine the permutation scheme and for this purpose it will be passed to genPerms and nrPerms.

asStat determines also the input data format if you declare your own stat type. In this case the two functions you write for computing the observed and permuted statistics will take the x and Y arguments as it follows.

For Stat belonging to group

BivStatistic : x is a fixed vector with $N$ elements and Y an an $M$-vector of $N$-vectors. The $M$ bivariate statistics between x and the $y_m$ vectors of Y are tested simultaneously. ns is ignored.

IndSampStatistic : we have $K$ groups and $N=N_1+...+N_K$ total observations; Y is an $M$-vector, each one holding all observations for the $m^{th}$ hypothesis in a single vector. The $m^{th}$ vector $y_m$ concatenates the observations for all groups such as [y[m][1];...;y[m][K]]. x is the membership(::IndSampStatistic) vector, common to all hypotheses. For example, for $K=2$, $N_1=2$ and $N_2=3$, x=[1, 1, 2, 2, 2].

RepMeasStatistic : we have $K$ measures (e.g., treatements) and $N$ subjects; Y is an M-vector, each one holding the $K*N$ observations for the $m^{th}$ hypothesis in a single vector. The $m^{th}$ vector $y_m$ is such as [Y[m][1];...;Y[m][N]], where each vector Y[1][n], for $n=1…N$, holds the observations for the $K$ treatments and x=collect(1:K*N) (see membership(::RepMeasStatistic)).

OneSampStatistic : We have $N$ observations (e.g., subjects); Y is an $M$-vector, each one holding the $N$ observations and x=ones(Int, N) (see membership(::OneSampStatistic)).

Optional keyword arguments switch2rand, nperm, standardized, centered, seed, fstat, compfunc and verbose have the same meaning as in the _permTest! function.

If threaded is true (default) the function is multi-threaded if the product of the number of hypotheses, observations, and permutations exceed 500 millions. If you have unexpected problems with the function, try setting threaded to false.

For the cpcd argument, see create you own test.

Return a MultcompTest structure.

The number of executed steps $S$ can be retrived as the length of the .rejections field of the returned structure.

Examples

using PermutationTests
N, M = 8, 100 # 100 hypotheses, N=8
x=randn(N)
Y=[randn(N) for m=1:M]
# bi-directional exact test of the correlation between x and 
# all the M vector in Y.
T12 = _permMcTest!(x, Y, N, PearsonR(), PearsonR())
T12.p
T12.stat
#...

# bi-directional exact test. Faster test by data standardization 
T12_2 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y in Y], N, CrossProd(), PearsonR(); 
                    standardized=true) 

# left-directional exact test.
T12_3 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y in Y], N, CrossProd(), PearsonR(); 
                    standardized=true, fstat=flip)

# as above, but force an approximate test
T12_4 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y in Y], N, CrossProd(), PearsonR(); 
                    standardized=true, fstat=flip, switch2rand=1)

# the same using 5000 random permutations
T12_4 = _permMcTest!(μ0σ1(x), [μ0σ1(y) for y in Y], N, CrossProd(), PearsonR(); 
                    standardized=true, fstat=flip, switch2rand=1, nperm=5000)

To check more examples, see the multcompTests_API.jl unit located in the src github folder and function test_multicompTests() in the runtests.jl unit located in the test github folder.

PermutationTests.flip — Function

flip(x::Bool)

flip(x::Union{R, I}) where {R<:Real, I<:Int}

Invert the sign of a real number or integer and negate a boolean.

This function is be needed only to call _permTest! or _permMcTest! if you create your own test.

PermutationTests.membership — Function

function membership(stat::IndSampStatistic, ns::Vector{Int})

Create the appropriate argument x to be used by functions _permTest! and _permMcTest! when you create your own test using the permutation scheme of test statistics belonging to the IndSampStatistic group.

For stat a IndSampStatistic, ns is a group numerosity vector, i.e., a vector of positive integers [N1,...,NK], where $K$ is the number of groups and $N_k$ is the number of observations for the $k^{th}$ group (see ns).

Return the group membership vector [repeat (1, N1);...; repeat(K, Nk)]

If rev=reverse is passed as keyword argument, return instead group membership vector [repeat (K, N1);...; repeat(1, Nk)]. This is used to run t-tests for independent samples using the PearsonR() Statistic, see for example studentTestIS.

Examples

using PermutationsTests
membership(AnovaF_IS(), [3, 4])
# return the vector [1, 1, 1, 2, 2, 2, 2]

function membership(stat::RepMeasStatistic, ns::@NamedTuple{n::Int, k::Int})

Create the appropriate argument x to be used by functions _permTest! and _permMcTest! when you create your own test using the permutation scheme of test statistics belonging to the RepMeasStatistic group.

For stat a RepMeasStatistic, ns is a named tuple, such as (n=N, k=K), where $N$ is the number of observations (e.g., subjects) and $K$ the number of measurements (or treatments, times, ect.), see ns.

Return collect(1:N*K).

Examples

using PermutationsTests
membership(AnovaF_RM(), (n=2, k=4))
# return the vector [1, 2, 3, 4, 5, 6, 7, 8]

function membership(stat::OneSampStatistic, ns::Int)

Create the appropriate argument x to be used by functions _permTest! and _permMcTest! when you create your own test using the permutation scheme of test statistics belonging to the OneSampStatistic group.

For stat a OneSampStatistic, ns is the number of observations (e.,g., subjects) given as an integer, see ns.

Return ones(Int, ns).

Examples

using PermutationsTests
membership(Sum(), 5)
# return the vector [1, 1, 1, 1, 1]

PermutationTests._observedStats — Function

function _observedStats(x, Y::UniDataVec, stat::statistic, fstat; pcd=nothing, kwargs...)
                        where statistic<:Statistic 

Compute obseved statistics for multiple comparisons permutation tests.

If you create your own test you will write a method taking as stat a test statistic of type Statistic you have declared.

If not, you never need this function.

PermutationTests._permutedStat — Function

function _permutedStat(x, Y::UniDataVec, i::Int, stat::AnovaF_RM, fstat; pcd=nothing, kwargs...) 

Compute the $i^{th}$ permuted statistics for the $i^{th}$ hypothesis, with $i=1...M$, for multiple comparisons permutation tests.

If you create your own test you will write a method taking as stat a test statistic of type Statistic you have declared.

If not, you never need this function.

PermutationTests._∑y² — Function

function _∑Y²(y)

Pre-compute some data for the statistic methods computing the StudentT_1S test-statistic and the AnovaF_RM test-statistic.

Return the sum of all squared observations in y.

PermutationTests._∑Y²kn_∑y²_∑S²k — Function

function _∑Y²kn_∑y²_∑S²k(y, ns)

Pre-compute some data for the statistic computing the AnovaF_RM test-statistic.

Return a 3-tuple.